service/requirements.txt
azure-storage-blob==12.8.0
pywin32==300
requests==2.25.1
